+++ src/systems/particles/index.js
@@ -1,0 +1,5 @@
+export { Particle } from "./Particle.js";
+export { ParticleManager } from "./ParticleManager.js";
+export { ParticleRenderer } from "./ParticleRenderer.js";
+export { ParticleState } from "./ParticleState.js";
+export { ParticleTrail } from "./ParticleTrail.js";